Relocate _read_varuint from base to plaintext framehelper
Only plaintext reads varuints on the wire, this should
be in plaintext and not base
