+++ aioesphomeapi/_frame_helper/base.py
@@ -125,22 +125,6 @@ class APIFrameHelper:
         # above to verify we never try to read past the end of the buffer
         return cstr[original_pos:new_pos]
 
-    def _read_varuint(self) -> _int:
-        """Read a varuint from the buffer or -1 if the buffer runs out of bytes."""
-        if TYPE_CHECKING:
-            assert self._buffer is not None, "Buffer should be set"
-        result = 0
-        bitpos = 0
-        cstr = self._buffer
-        while self._buffer_len > self._pos:
-            val = cstr[self._pos]
-            self._pos += 1
-            result |= (val & 0x7F) << bitpos
-            if (val & 0x80) == 0:
-                return result
-            bitpos += 7
-        return -1
-
     @abstractmethod
     def write_packets(
         self, packets: list[tuple[int, bytes]], debug_enabled: bool
